# client/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eyened API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/studies/{study_id}/tags": {
      "post": {
        "summary": "Tag Study",
        "description": "Attach a Tag to a Study by tag ID (idempotent).",
        "operationId": "tag_study_studies__study_id__tags_post",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPOST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/studies/{study_id}/tags/{tag_id}": {
      "delete": {
        "summary": "Untag Study",
        "description": "Remove a Tag from a Study (idempotent).",
        "operationId": "untag_study_studies__study_id__tags__tag_id__delete",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Study Tag",
        "description": "Update comment on an existing Study tag link.",
        "operationId": "patch_study_tag_studies__study_id__tags__tag_id__patch",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "summary": "Get Patient",
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patient Id"
            }
          },
          {
            "name": "include_attributes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Attributes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDetailGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatorMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CreatorMeta"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version"
        ],
        "title": "ModelMeta"
      },
...
      "ObjectTagPATCH": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "ObjectTagPATCH"
      },
      "ObjectTagPOST": {
        "properties": {
          "tag_id": {
            "type": "integer",
            "title": "Tag Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "tag_id"
        ],
        "title": "ObjectTagPOST"
      },
      "PatientAttributeValueGET": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "PatientAttributeValueGET",
        "description": "One attribute value produced by a model (or manual entry when model is null)."
      },
      "PatientDetailGET": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SexEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "project": {
            "$ref": "#/components/schemas/ProjectMeta"
          },
          "attrs": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/PatientAttributeValueGET"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Attrs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identifier",
          "project"
        ],
        "title": "PatientDetailGET"
      },
...
      "ProjectMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectMeta"
      },
...
      "SexEnum": {
        "type": "string",
        "enum": [
          "M",
          "F"
        ],
        "title": "SexEnum"
      },
...
      "TagMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tagger": {
            "$ref": "#/components/schemas/CreatorMeta"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "tagger",
          "date"
        ],
        "title": "TagMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}